Client/requirements.txt
cryptography==36.0.2
Flask==2.1.2
Flask_Cors==3.0.10
pycryptodome==3.14.1
PyJWT==2.3.0
bcrypt==3.2.0
requests==2.27.1
